DjangoApplication/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-auditlog==3.0.0
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.6
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
torch
torchvision
